net: ndo_fdb_dump int* -> u64*
Note: I am leaving this patch in for the RFC, but only because I think the
      RB tree approach in the next patch is a non-starter, and I don't want
      to waste time on cleaning this up. I elaborate in the next patch.
      Feel free to skip. The tl;dr is that we use (ETH,VID) tuple for the
      index.

This NDO uses an int* as an out-parameter to save progress along the list
of FDB entries to dump. In the following patch, the bridge driver will
convert to using a red-black tree to organize the FDB in a way that is more
suitable for dump restarts. Now if we sort the tree according to the (ETH
address, VID) tuple, which is 8 bytes, and if the ndo_fdb_dump NDO used u64
indices, we could just use the tuple as the index itself.

In this patch, change ndo_fdb_dump API to use u64 for index to support the
scheme described above.

+++ drivers/net/ethernet/freescale/dpaa2/dpaa2-switch.c
@@ -903,7 +903,7 @@ static int dpaa2_switch_fdb_entry_dump(struct ethsw_port_priv *port_priv,
 
 static int dpaa2_switch_port_fdb_dump(struct sk_buff *skb, struct netlink_callback *cb,
 				      struct net_device *net_dev,
-				      struct net_device *filter_dev, int *idx)
+				      struct net_device *filter_dev, u64 *idx)
 {
 	struct ethsw_port_priv *port_priv = netdev_priv(net_dev);
 	struct ethsw_dump_ctx dump = {
